typescript/requirements.txt
httpx
pyyaml
click
rich
openai
